+++ src/frontend/komponenter/Fagsak/Vedtak/VedtakBegrunnelserTabell/AvslagBegrunnelser.tsx
@@ -4,15 +4,21 @@ import { useVedtakBegrunnelser } from '../../../../context/VedtakBegrunnelseCont
 import { IBehandling } from '../../../../typer/behandling';
 import { IRestAvslagbegrunnelser } from '../../../../typer/vedtak';
 import { RessursStatus } from '@navikt/familie-typer';
-import AvslagBegrunnelsePanel from './AvslagBegrunnelsePanel';
 import { Vedtaksperiode, Vedtaksperiodetype } from '../../../../typer/vedtaksperiode';
 import familieDayjs, { familieDayjsDiff } from '../../../../utils/familieDayjs';
 import { datoformat } from '../../../../utils/formatter';
-import OverskriftMedHjelpetekst from './OverskriftMedHjelpetekst';
+import OverskriftMedHjelpetekst from './Felles/OverskriftMedHjelpetekst';
+import EkspanderbartBegrunnelsePanel from './Felles/EkspanderbartBegrunnelsePanel';
+import { Normaltekst } from 'nav-frontend-typografi';
+import styled from 'styled-components';
 
 interface IAvslagTabell {
     åpenBehandling: IBehandling;
 }
+
+const MarginDiv = styled.div`
+    margin-left: 0.625rem;
+`;
 
 const AvslagBegrunnelser: React.FC<IAvslagTabell> = ({ åpenBehandling }) => {
     const { avslagBegrunnelser } = useVedtakBegrunnelser();
@@ -39,16 +45,19 @@ const AvslagBegrunnelser: React.FC<IAvslagTabell> = ({ åpenBehandling }) => {
                           )
                 )
                 .map((periode: Vedtaksperiode) => (
-                    <AvslagBegrunnelsePanel
-                        vedtaksperiode={periode}
-                        begrunnelser={
-                            avslagBegrunnelser.data.find(
-                                (avslagBegrunnelser: IRestAvslagbegrunnelser) =>
-                                    avslagBegrunnelser.fom === periode.periodeFom &&
-                                    avslagBegrunnelser.tom === periode.periodeTom
-                            )?.brevBegrunnelser ?? []
-                        }
-                    />
+                    <EkspanderbartBegrunnelsePanel vedtaksperiode={periode} åpen={true}>
+                        <MarginDiv>
+                            <Normaltekst children={'Begrunnelse(r) for avslag'} />
+                            <ul>
+                                {avslagBegrunnelser.data.find(
+                                    (avslagBegrunnelser: IRestAvslagbegrunnelser) =>
+                                        avslagBegrunnelser.fom === periode.periodeFom &&
+                                        avslagBegrunnelser.tom === periode.periodeTom
+                                )?.brevBegrunnelser ??
+                                    [].map((begrunnelse: string) => <li>{begrunnelse}</li>)}
+                            </ul>
+                        </MarginDiv>
+                    </EkspanderbartBegrunnelsePanel>
                 ))}
         </>
     ) : null;
