+++ src/frontend/komponenter/Fagsak/Vedtak/VedtakBegrunnelserTabell/VedtakBegrunnelser.tsx
@@ -3,19 +3,47 @@ import React from 'react';
 import { useApp } from '../../../../context/AppContext';
 import { useBehandling } from '../../../../context/BehandlingContext';
 import { useVedtakBegrunnelser } from '../../../../context/VedtakBegrunnelseContext';
-import { IBehandling } from '../../../../typer/behandling';
-import { lagPeriodeId } from '../../../../typer/periode';
+import { Behandlingstype, IBehandling } from '../../../../typer/behandling';
 import { ToggleNavn } from '../../../../typer/toggles';
 import { IRestVedtakBegrunnelse } from '../../../../typer/vedtak';
-import { Vedtaksperiode, Vedtaksperiodetype } from '../../../../typer/vedtaksperiode';
+import {
+    IUtbetalingsperiodeDetalj,
+    Vedtaksperiode,
+    Vedtaksperiodetype,
+} from '../../../../typer/vedtaksperiode';
 import familieDayjs, { familieDayjsDiff } from '../../../../utils/familieDayjs';
-import { datoformat } from '../../../../utils/formatter';
-import VedtakBegrunnelsePanel from './VedtaksBegrunnelsePanel';
-import OverskriftMedHjelpetekst from './OverskriftMedHjelpetekst';
+import { datoformat, formaterBeløp, formaterPersonIdent } from '../../../../utils/formatter';
+import OverskriftMedHjelpetekst from './Felles/OverskriftMedHjelpetekst';
+import { Element, Normaltekst } from 'nav-frontend-typografi';
+import VedtakBegrunnelserMultiselect from './VedtakBegrunnelserMultiselect';
+import EkspanderbartBegrunnelsePanel from './Felles/EkspanderbartBegrunnelsePanel';
+import { IRestPersonResultat } from '../../../../typer/vilkår';
+import styled from 'styled-components';
 
 interface IVedtakBegrunnelserTabell {
     åpenBehandling: IBehandling;
 }
+
+interface IVedtakBegrunnelserTabell {
+    vedtaksperiode: Vedtaksperiode;
+    personResultater: IRestPersonResultat[];
+    behandlingsType: Behandlingstype;
+}
+
+const UtbetalingsperiodepanelBody = styled.div`
+    margin-left: 0.625rem;
+    display: grid;
+    grid-template-columns: 5fr 4fr;
+`;
+
+const UtbetalingsperiodeDetalj = styled.div`
+    display: flex;
+    flex-direction: row;
+
+    .typo-normal {
+        margin-right: 1.5rem;
+    }
+`;
 
 const VedtakBegrunnelser: React.FC<IVedtakBegrunnelserTabell> = ({ åpenBehandling }) => {
     const { toggles } = useApp();
@@ -73,15 +101,43 @@ const VedtakBegrunnelser: React.FC<IVedtakBegrunnelserTabell> = ({ åpenBehandli
                     }
                 })
                 .map((vedtaksperiode: Vedtaksperiode) => (
-                    <VedtakBegrunnelsePanel
-                        key={lagPeriodeId({
-                            fom: vedtaksperiode.periodeFom,
-                            tom: vedtaksperiode.periodeTom,
-                        })}
-                        behandlingsType={åpenBehandling.type}
-                        personResultater={åpenBehandling.personResultater}
+                    <EkspanderbartBegrunnelsePanel
                         vedtaksperiode={vedtaksperiode}
-                    />
+                        åpen={åpenBehandling.type === Behandlingstype.FØRSTEGANGSBEHANDLING}
+                    >
+                        <UtbetalingsperiodepanelBody>
+                            {vedtaksperiode.vedtaksperiodetype === Vedtaksperiodetype.UTBETALING ? (
+                                <div>
+                                    <Element>Resultat</Element>
+
+                                    {vedtaksperiode.utbetalingsperiodeDetaljer.map(
+                                        (detalj: IUtbetalingsperiodeDetalj) => (
+                                            <UtbetalingsperiodeDetalj
+                                                key={detalj.person.personIdent}
+                                            >
+                                                <Normaltekst title={detalj.person.navn}>
+                                                    {formaterPersonIdent(detalj.person.personIdent)}
+                                                </Normaltekst>
+
+                                                <Normaltekst>
+                                                    {formaterBeløp(detalj.utbetaltPerMnd)}
+                                                </Normaltekst>
+                                            </UtbetalingsperiodeDetalj>
+                                        )
+                                    )}
+                                </div>
+                            ) : (
+                                <div />
+                            )}
+                            <div>
+                                <VedtakBegrunnelserMultiselect
+                                    erLesevisning={erLesevisning()}
+                                    personResultater={åpenBehandling.personResultater}
+                                    vedtaksperiode={vedtaksperiode}
+                                />
+                            </div>
+                        </UtbetalingsperiodepanelBody>
+                    </EkspanderbartBegrunnelsePanel>
                 ))}
         </>
     ) : null;
